Create a Policy for each backend. Update events to use new Policy
This commit adds a new Policy class which will be used to create a similar API
for each backend. This commit also updates the Event object to use the new
Policy class instead of the old EventPolicy class.

This class will work on low level functions for each backend and work on
handles. This class will be used by other classes to provide the logic
for each backend without repeating logic.